feat(coins): LS-1302 show balance
