+++ src/styles/index.scss
@@ -19,4 +19,3 @@
 @import '../styles/components/curveSwitch.scss';
 @import '../components/LSNumberInput/lsNumberInputAutoNumeric.scss';
 @import '../Swap/styles/swapInput';
-
