+++ src/styles/modals.scss
@@ -12,4 +12,5 @@
 @import '../components/InvalidNetworkDialog/invalidNetworkDialog';
 @import '../components/SelectTokenDialog/managePresets';
 @import '../components/ContractSwitch/contractSwitch';
+@import '../components/SelectTokenDialog/tokenList';
